Restructure sources and templates by framework
