+++ build2cmake/src/templates/cuda/_ops.py
