+++ build2cmake/src/templates/cuda/hipify.py
