+++ build2cmake/src/templates/cuda/kernel.cmake
