+++ build2cmake/src/templates/cuda/preamble.cmake
