+++ build2cmake/src/templates/cuda/registration.h
