+++ build2cmake/src/templates/cuda/setup.py
