+++ build2cmake/src/templates/cuda/utils.cmake
