+++ build2cmake/src/templates/universal/_ops.py
