+++ build2cmake/src/templates/universal/pyproject.toml
