+++ build2cmake/src/torch/cuda.rs
@@ -1,52 +1,19 @@
 use std::collections::HashSet;
 use std::io::Write;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
 use eyre::{bail, Context, Result};
-use git2::Repository;
 use itertools::Itertools;
 use minijinja::{context, Environment};
-use rand::Rng;
 
+use super::kernel_ops_identifier;
 use crate::config::{Backend, Build, Dependencies, Kernel, Torch};
 use crate::FileSet;
 
-static CMAKE_UTILS: &str = include_str!("cmake/utils.cmake");
-static REGISTRATION_H: &str = include_str!("templates/registration.h");
-static HIPIFY: &str = include_str!("cmake/hipify.py");
-static CUDA_SUPPORTED_ARCHS_JSON: &str = include_str!("cuda_supported_archs.json");
-
-fn random_identifier() -> String {
-    // Generate a random string when no ops_id is provided
-    let mut rng = rand::thread_rng();
-    let build_id: u64 = rng.gen();
-    base32::encode(
-        base32::Alphabet::Rfc4648Lower { padding: false },
-        &build_id.to_le_bytes(),
-    )
-}
-
-fn git_identifier(target_dir: impl AsRef<Path>) -> Result<String> {
-    let repo = Repository::discover(target_dir.as_ref()).context("Cannot open git repository")?;
-    let head = repo.head()?;
-    let commit = head.peel_to_commit()?;
-    let rev = commit.tree_id().to_string().chars().take(7).collect();
-    let dirty = !repo.statuses(None)?.is_empty();
-    Ok(if dirty { format!("{rev}_dirty") } else { rev })
-}
-
-pub fn kernel_ops_identifier(
-    target_dir: impl AsRef<Path>,
-    name: &str,
-    ops_id: Option<String>,
-) -> String {
-    let identifier = ops_id.unwrap_or_else(|| match git_identifier(target_dir.as_ref()) {
-        Ok(rev) => rev,
-        Err(_) => random_identifier(),
-    });
-
-    format!("_{name}_{identifier}")
-}
+static CMAKE_UTILS: &str = include_str!("../templates/cuda/utils.cmake");
+static REGISTRATION_H: &str = include_str!("../templates/cuda/registration.h");
+static HIPIFY: &str = include_str!("../templates/cuda/hipify.py");
+static CUDA_SUPPORTED_ARCHS_JSON: &str = include_str!("../cuda_supported_archs.json");
 
 fn cuda_supported_archs() -> String {
     let supported_archs: Vec<String> = serde_json::from_str(CUDA_SUPPORTED_ARCHS_JSON)
@@ -131,7 +98,7 @@ fn write_setup_py(
 
     let data_globs = torch.data_globs().map(|globs| globs.join(", "));
 
-    env.get_template("setup.py")
+    env.get_template("cuda/setup.py")
         .wrap_err("Cannot get setup.py template")?
         .render_to_write(
             context! {
@@ -159,7 +126,7 @@ fn write_ops_py(
     path.push("_ops.py");
     let writer = file_set.entry(path);
 
-    env.get_template("_ops.py")
+    env.get_template("cuda/_ops.py")
         .wrap_err("Cannot get _ops.py template")?
         .render_to_write(
             context! {
@@ -217,7 +184,7 @@ pub fn render_binding(
     name: &str,
     write: &mut impl Write,
 ) -> Result<()> {
-    env.get_template("torch-binding.cmake")
+    env.get_template("cuda/torch-binding.cmake")
         .wrap_err("Cannot get Torch binding template")?
         .render_to_write(
             context! {
@@ -243,7 +210,7 @@ fn render_deps(env: &Environment, build: &Build, write: &mut impl Write) -> Resu
     for dep in deps {
         match dep {
             Dependencies::Cutlass2_10 => {
-                env.get_template("dep-cutlass.cmake")
+                env.get_template("cuda/dep-cutlass.cmake")
                     .wrap_err("Cannot get CUTLASS dependency template")?
                     .render_to_write(
                         context! {
@@ -254,7 +221,7 @@ fn render_deps(env: &Environment, build: &Build, write: &mut impl Write) -> Resu
                     .wrap_err("Cannot render CUTLASS dependency template")?;
             }
             Dependencies::Cutlass3_5 => {
-                env.get_template("dep-cutlass.cmake")
+                env.get_template("cuda/dep-cutlass.cmake")
                     .wrap_err("Cannot get CUTLASS dependency template")?
                     .render_to_write(
                         context! {
@@ -265,7 +232,7 @@ fn render_deps(env: &Environment, build: &Build, write: &mut impl Write) -> Resu
                     .wrap_err("Cannot render CUTLASS dependency template")?;
             }
             Dependencies::Cutlass3_6 => {
-                env.get_template("dep-cutlass.cmake")
+                env.get_template("cuda/dep-cutlass.cmake")
                     .wrap_err("Cannot get CUTLASS dependency template")?
                     .render_to_write(
                         context! {
@@ -276,7 +243,7 @@ fn render_deps(env: &Environment, build: &Build, write: &mut impl Write) -> Resu
                     .wrap_err("Cannot render CUTLASS dependency template")?;
             }
             Dependencies::Cutlass3_8 => {
-                env.get_template("dep-cutlass.cmake")
+                env.get_template("cuda/dep-cutlass.cmake")
                     .wrap_err("Cannot get CUTLASS dependency template")?
                     .render_to_write(
                         context! {
@@ -308,7 +275,7 @@ pub fn render_kernel(
         .collect_vec()
         .join("\n");
 
-    env.get_template("kernel.cmake")
+    env.get_template("cuda/kernel.cmake")
         .wrap_err("Cannot get kernel template")?
         .render_to_write(
             context! {
@@ -329,7 +296,7 @@ pub fn render_kernel(
 }
 
 pub fn render_extension(env: &Environment, ops_name: &str, write: &mut impl Write) -> Result<()> {
-    env.get_template("torch-extension.cmake")
+    env.get_template("cuda/torch-extension.cmake")
         .wrap_err("Cannot get Torch extension template")?
         .render_to_write(
             context! {
@@ -345,7 +312,7 @@ pub fn render_extension(env: &Environment, ops_name: &str, write: &mut impl Writ
 }
 
 pub fn render_preamble(env: &Environment, name: &str, write: &mut impl Write) -> Result<()> {
-    env.get_template("preamble.cmake")
+    env.get_template("cuda/preamble.cmake")
         .wrap_err("Cannot get CMake prelude template")?
         .render_to_write(
             context! {
